// api/src/domain/sabotage-card/sabotage-card-names.ts
export enum SABOTAGE_CARD_NAMES {
  PLANO_COHEN = 'PLANO_COHEN',
  MENSALAO = 'MENSALAO',
  CAFE_COM_A_ABIN = 'CAFE_COM_A_ABIN',
  // OPERACAO_MAOS_LIMPAS = "OPERACAO_MAOS_LIMPAS",
  O_FMI_MANDOU = 'O_FMI_MANDOU',
  FORCAS_OCULTAS = 'FORCAS_OCULTAS',
  // REGIME_DE_URGENCIA = "REGIME_DE_URGENCIA",
  SESSAO_SECRETA = 'SESSAO_SECRETA',
  GOLPE_DE_ESTADO = 'GOLPE_DE_ESTADO',
  // VAZAMENTO_NO_WIKILEAKS = "VAZAMENTO_NO_WIKILEAKS",
  CONGRESSO_TRANCADO = 'CONGRESSO_TRANCADO',
  PEGADINHA_DO_PARAGRAFO_47_INCISO_V = 'PEGADINHA_DO_PARAGRAFO_47_INCISO_V',
  VETO_DO_STF = 'VETO_DO_STF',
  // DELACAO_PREMIADA = "DELACAO_PREMIADA",
  // MENSAGEM_ANONIMA = "MENSAGEM_ANONIMA",
  // TUITACO = "TUITACO",
  // RECONTAGEM_VOTOS = "RECONTAGEM_VOTOS",
  // CENSURA_ESTATAL = "CENSURA_ESTATAL",
  // PACOTE_DE_LEIS = "PACOTE_DE_LEIS",
  // VOTO_DE_MINERVA = "VOTO_DE_MINERVA",
  // FRAUDE_ELEITORAL = "FRAUDE_ELEITORAL",
}